Make it possible to handle feedback from the server.
This is now possible because we track starting and ending times of
merges within the core. This is possible because the merger has to
return the merge time on a `merge()' method call. When calling
`get_user_feedback()' method call on the communicator the start and end
time of the feedback is also given. Al this behavior is also tested. It
is not yet integrated with the picker to actually do something with this
feedback.