Make NCAPicker reentrant save if the force flag is specified.
Because it can happen that the merger cannot merge two songs, it can
also happen that a new song has to be chosen before we goto the next
part. This commit makes sure this doesn't affect the feedback chosing as
it could happen that these two get out of sync. This commit also
includes tests to make sure the feedback state is not changed when
called `get_next_song()' is called with `force=True'.