Fix a bug in the tests and with parsing type information
A bug in the tests was found that lead to not checking if all params
where parsed correctly by `parse_docstring` from helpers.py. This is
fixed as well as a bug that was found because of this improved tests.
This bug happened if the users provided type information which would get
parsed as a part of the variable name. This bug is also included in the
tests.