Create a first version of the start command
This includes a working version of the consumer producer setup in web.py
that uses two processes: one to process incomming request and one to
process the actual commands. These are connected by a queue.

The core also does a started response just after the first sample is
created as per specification.