Fixed logic error on non empty queue start call
If the queue was not empty when we receive the start call this means
that the first item would be become the last item of the queue. However
we could have a queue of `[ADD song1, ADD song2, DELETE song1]` which
would then become `[ADD song2, DELETE song1, ADD song1]` which would
mean that `song1` would be added in the end.