Fix a bug where the picker can get stuck in a force loop
This is because it was filtering songs, however it can happen that this
leaves only songs which are simply not suitable for the transitioner.
The NCAPicker now keeps track of a `force_streak` variable and after 2
calls to `get_next_song` with `force=True` without a call to
`get_next_method` with `force=False` it will stop filtering based on
tempo.